[C++] Fix message display functions re. versioning.
As coded it wouldn't handle messages from previous schema versions if the block
length had grown.

The part I removed feels like it was put there deliberately to always use the
block length from the current codec rather than the block length in the message.
But that also appears to be the root of the problem itself.